Greatly improve composability with other languages.
Namely, a new #lang can add editor support out of the box without having
to rely on all files to start with #lang editor.

Also, the lexer now mixes with other lexers.

And finally, langs that require front matter to appear before any other
code can request it with the `#:headers` field in the editor's
`#%module-begin`.