BREAKING CHANGE!!! More robust editor deserialization.
Whenever a new editor was added into the hiarchy, it would break all
editors that extended it. This runs a symbolic check to make a best
effort attempt to properly deserialize the editor.

Also add a 'version' field to editor (de)serialization to hopefully
reduce the amount of breaking changes.